refactor: getTotalCount
